Add the interface to create new local playlists
Add a `.navbar` element to the local playlist's page with a link to
create new playlists. The link simply shows up a bubble with a form to
fill in the title of the new playlist and once submitted, the playlist
is added to the database and we display a notification to the user.

We do not directly delegate to the local model since we may implement
this feature for other services in the future so the delegation is done
inside the meta model.