Unify the downloading interface
The previous downloading interface was really not well-designed and
handled. We used to display a progress bar on the element that was
downloaded but the user can't really see the progress easily nor check
for the chosen tags, etc.

So let's introduce a `Downloads` object that will handle our downloading
queue and the download history. These elements are displayed on the
meta/downloads action so the user can see the downloading elements and
their history clicking on the dashboard icon.

The dashboard icon also now has a download progress bar so the user can
see the progress of the whole download queue (not only of one element
because the downloads can happen simultaneously).

Also introduce a new `.download.box` element to display downloading or
downloaded elements on the `meta/downloads` action.

Finally, split the Ui downloading function into several functions
(namely downloadStart, downloadProgress and downloadEnd) to make
things clearer and avoid passing arguments when they are not needed.