Rely on the new downloading interface
Use the new downloading interface inside the SoundCloud and YouTube
services when downloading medias.

Also avoid passing the album rarameter for now to the meta service
`downloadImage` function for now as we are not proposing the user to
set the album for now (this is a remain of an old try that has never
been finished).