Remove the shadow.js event handler file
The name of this file didn't really told its intent; for sure the video
player is inside the `.shadow` element but it's not that obvious that
the event handlers of the video player are stored inside this file.

So let's move the only `.shadow` event handler to the container's event
handlers and create a file dedicated to the video player's event
handlers.