Ensure the interface works as expected changing the locale
When the user picks a different locale inside the configuration page and
saves the configuration, we are calling `Ui#loadPartials`. The problem is
that this function erases the HTML content of the different sections of
the interface when the new HTML is loaded.

Thus, we need to make sure that the event handlers are binded to an
element that won't be removed by this function (here, `.sidebar`).

Also, this function loads the partials for the player controls so we
need to pass it a boolean that will avoid reloading these partials. Thus
we avoid extra computation (because these partials are not relying on
any translation) and if the user launched a media before changing the
preferred locale, the player bar and video controls will still work as
expected.