Correctly show the controls in full-screen mode
When switching a video to the full-screen mode, let's display the play-
pause button and the timing information and controls so the users won't
have to switch to the normal mode if they wants to access to one of
these.

Also rename the `.player_frame_container` div to `.video.player` because
this is just what it is, the previous name wasn't telling anything about
the element and also now we can share some CSS rules and JavaScript
event handlers between the main player bar and this element.

Finally remove the download and share icons from the video player since
they are not really useful and the user can still realize these actions
quitting the video player.