Improve artist guessing
When downloading a music, we are trying to infer the artist. The
problem was that we were maintaining a list of labels and checking
whether the account was a label or not to decide whether to use it or
not.

However, this solution is not viable so let's rather guess the author
based on the title first (e.g. if there's "Remix" inside of it) and then
fall-back to the given account if we can't pick the artist from title.