Make the code clonable by anyone
Avoid referencing a folder on our file system and rather rely on the Git
repository for the tagging library.

Also let's avoid assuming that the `node_modules` folder is in the
parent directory of the repository to require the Less asset.

Finally, add a `rebuild` script to rebuild our tagging library to make
sure that there won't be any V8 version mismatch.