Correctly reset the active sidebar link
Clicking a link inside the sidebar should remove the active class of the
previous clicked link and add that class to the clicked one unless it is
an icon (i.e. one of home, search or dahsboard).

We should also special case the "search icon" as it is a link but which
is not displaying any new action so we don't have to reset the active
state.