Add a page for showing artists' works
Add a local/artist action that displays all the work of an artist
like their albums and singles.

For now, this is very basic but in the future we can widely improve
this to display all the work coming from the user's computer, SoundCloud
and YouTube.

The `div.artist` elements are now `a` elements to link to the artist's
page.