Improve a bit the playlist thumbnails styling
Set a fixed height for playlist thumbnails' images since images are not
loaded directly but the YouTube API is very fast to load them so we the
`dataShouldBeLoaded` function may return true since there's still
available space in the view-port while it should not.

Also reduce a bit the margins to have a more consistent styles with
other sections.