Properly change the DOM tagging a local music
When tagging a local music, we need to reflect the user's change inside
the DOM since audio files won't be re-processed even though they changed
on disk.

The problem is that if the user adds an artist field to the file while
it was empty before, we need to properly update the link's attributes so
that clicking on the artist's name will display the artist's page.

Let's also properly update the `title` attribute for the `.title` and
`.artist` elements.