Allow full qualifiers for SoundCloud and YouTube translations
SoundCloud and YouTube translations used to be only available under a
short-handed namespace (i.e. 'sc' and 'yt' respectively).

The problem is that, even though this is neat inside views, we need an
awkward mapping sometimes inside JavaScript files to access a
translation key under these sections.

Now, it's possible to use both `soundcloud.key` and `sc.key` to access a
SoundCloud translation for instance.