Remove some extra mark-up for playlist options
We don't need to specify the playlists' service since as of a3be1c16,
we are no longer displaying the playlists' thumbnails.

Also let's avoid relying on the `li.music` and `li.video` styles to give
a style to the playlist options so we have a consistent style across
services and we can customize them in the future if we want.