Improve consistency to manage services collection
Now, all collection are correctly concatenated inside the Cache class
rather than relying on the meta model to infer the model to use to
concatenate the existing collection with the brand new one.

Therefore, we are now consistent with the field names between SoundCloud
and YouTube. Both use `next_token` to refer to the pagination parameter
rather than `next_href` or `page_token` and both use `collection` to
refer to the array containing all the data.

This is not yet perfect though.