Improve record linking
Remove this closure from the meta model and put it under the Record
scope.

Also, move the linking calls to the cache since this is where data are
concatenated and are in a sort of "final state"; this avoid a lot of
calls when we are fetching data.

Finally, we now rely on `forEach` rather than `map` to run this closure
since we don't need to modify the array, we are just iterating over each
element to link it to the nearest ones. This saves us from *a bunch* of
allocations.