Fix reset of the YouTube history and "Listen later" at boot
Each time we boot the application, the YouTube history and "Listen
later" playlist files were reset since we mistakenly added a semi-colon
just next to the `if` line so the `touchPlaylistFile` call was made no
matter the result of the if.

Moral of the story: it's not so great to avoid curly braces around ifs.
:trollface: