Widely improve our model stack
Rather than only dealing with a `Record` object that can handle either
an album, or a playlist, or a media or a SoundCloud activity, let's
rather split all of these different possibilities into different
objects.

This brings some advantages:

* We are using less memory as we are not defining fields that may be
  useless in some cases (e.g. the `origin` field that is only useful for
  SoundCloud activities).
* The latter point also works for readability as having a `origin` or
  `type` field on something else that an activity wouldn't make any
  sense.
* The code is more clear inside the different service's method as we
  avoid a lot of checks to infer the kind of the given hash.

The changes bring to the YouTube client seem unrelated but before we
were wrapping results inside a `Record` object which is far from ideal;
now we are creating a hash picking the attributes that we need inside
the model to put it inside the returned Promise like in the SoundCloud
model. Thus, we can match the field name of the YouTube API to be
consistent.