Use the new `set` interface on medias inside Player
The Player class now relies on the fact that media instances have a
`set` attribute when they are wrapped inside a playlist or an album.

Also use a new terminology inside the player's markup to reflect the
fact that we are dealing with a set and not specifically a playlist.