Improve (again) our search feature
Let's remove the switches from the search options. They aren't nice at
all and don't look good with the rest of the user interface.

Now, on every search results pages we are displaying a tab to go to each
of the services so that a same query is run as late as possible rather
than querying all the picked services before displaying even one of
them.

Also, with this "new" interface, the users don't have to click on the
search options button and pick another service if they want to; it's
far easier now.