Ensure that the download folder exists before downloading
If the user tries to download a media for the first time or if the
location was valid before but has been removed or simply moved, we
need to tell the user that another path should be selected.

This will avoid errors and avoid downloading the file for having
nothing at the end.

We are now relying on promises to know when to trigger the specific
service action.

This prevents us from defining awkward event handlers and avoid a sort
of useless round-tripping because we first know the DOM element that has
been chosen and we convert it to a record but when the event handler is
triggered, we only have access to the record's id to find the original
DOM element.

This is also far easier to add other scenarios or error handling.