Avoid directly referencing service classes
Let's rather rely on our new `for` API to reference service classes.

Also, the YouTubeService#decrypt method isn't polluting the class'
scope anymore ; we still need to store the function's inner object
somewhere so the cache seems adequate enough but this is because
evaluating a raw string containing an object with functions isn't
allowed in ES6.