Properly handle playing singles
Artists' singles used to reference records stored inside the local
singles collection. The problem is that the next and previous records
weren't referencing the right once (at least from the user's point of
view).

Thus, even though this takes a bit more memory, there should be a
distinct collection for singles and a distinct `singles` collection
for each artist.

Finally, finding a record to play is now handled a bit differently
whether the user is on an artist's page or on the local/singles action.