Widely improve the user interface
Rely on the `hidden-inset` value of the `titleBarStyle` window attribute
to display a better title bar.

The navbar that were presents on certain sections are now part of the
title bar as links. The Ui.Menu.define function automatically infers
whether a menu should be added to the title bar or not.

Also make the search bar always visible. This avoids too much click and
in the future we can imagine different features like live filtering for
the current action's items.

Finally, the partials now live under the `app/views` folder as they
aren't part of the meta section at all. The title bar markup lives in
its own view.