Bring back the search feature
It was almost disabled as there were some remains (mostly the `source`
argument) of b8c0735 so it couldn't be triggered.

Now, we are no longer allowing searches inside the user's library. We
are directly hitting SoundCloud or YouTube hitting enter.

In the future, we may implement "live sorting" when the user types
inside the search bar so this would mimic the previous behavior.

Finally, the search by artist is disabled as we are now displaying the
local library sorted by artists so the user can type an artist's name
and every artist matching with the request will be displayed on the
search results page rather than displaying all of their medias.