Avoid caching data inside promises
Since we only want to wrap data within a promise when it is accessed
from the model, let's rather store it roughly for easier reuses when
we know that the data is already computed.

The `fetch` method has been added as a convenient way to wrap cached
data inside a Promise object.

Thus, we can easily fetch the current context even for paginated
sections as picking the view's context would only give the new data
loading the next page.