Give access to medias' album's title
Since we are wrapping album items under album objects but since they
are internally considered as set, the album's title wasn't properly
displayed inside the tagging view.

Actually, media instances used to have an `album` field which was only
containing the title but since we are storing data in a very different
way, accessing this field was no longer valid.