Allow renaming artists by double-clicking their name
Sometimes, a distinction could be make between two artists inside
DaPlayer while it's only a single one (e.g. Rise of the North Star
and Rise of the Northstar).

We could add some guards to avoid such problem but this would make
the code a bit more complicated and we can't handle all the possible
cases so let's rather allow the user to rename artists and warn them
when the name is already taken.