Change the connection handling for the YouTube section
We should clearly make the distinction between the fact that an
account that has been connected with the application and the fact
that an account has a valid token to pull data from the API.

Thus, since we make this distinction now, we can return a Promise
trying to refresh the user's token and make sure that we won't have
a random failure when the application has been running for a long time
and the token is no longer valid.

This hardens our code and simplifies it, avoiding a lot of Promises
allocations.