Fix item loading for reposted SoundCloud playlists
If the activity feed contains some playlists, previously, the user
could have seen some missing items.

This is because SoundCloud will fetch items for reposted playlists
but as many as the `limit` query parameter specifies (which is used
to ask for a certain number of activities, not items in playlists).

Thus, we need to make sure to fetch the items for playlists that
have more than 10 items loading the feed.