Don't try to eagerly connect the YouTube account
We use to automatically refresh the user's token on boot (there is
no trace of this in the git history, just legacy) since the token
was most-likely no longer valid.

However, now the `YT.fetch` method properly refreshes the token when
pulling data if it's needed. Thus, let's rather avoid a potentially
useless operation and speed up a bit the boot time.