Widely improve the displaying of the SoundCloud stream
Change the height of music bars to make the view display a bit more
element and in order to stick with SoundCloud's styling.

Also improve a bit the activity description by using colors that
are more consistent with our interface and improve the image displaying
with a tiny border.

Finally, now, the `.music.bar` elements displays the playback and
repost counts and some action buttons just like any other media
elements.