Widely improve the Context API
The `Context` constructor used to do a pretty tricky dance as it
had to sort of infer the keys where data was. Apart from the fact
that we could've passed that key in the first place, this is actually
no longer needed.

Since as of ab94eaaf, all our response are consistent and have a
`collection` key (apart from page for `Artist`, `Playlist` and
`Array`), we can safely assume that this key exist and build a
flattened collection on top of this one.

This has several advantages:

* We are reducing the number of allocations on a per action basis.
* The `Context` constructor is a lot simpler.
* It is super easy to check whether we are dealing with a specific
  kind of record.

This last point allows us to fix a pretty long-lasting bug that
prevent people from playing playlist items on the SoundCloud stream
and also fix any method relying on `Record.from` as the latter pull
data from the current context and can thus sometimes return an
instance of the `Activity` class.