Avoid using links when we aren't rendering anything
Let's be semantic and use the proper kind of DOM elements. The tab
menu for dialogs shouldn't use `<a>` elements as they are used to
render actions throughout the application.

Also, the button used to pick another picture tagging a media isn't
rendering anything either so let's use a span instead of a link.