Fix the YouTube sign-in logic
Now, YouTube is automatically injecting the content of the page
rather than doing actual requests during the authorization process
so we need to listen to a different event to check whether we are
on the final page with the authorization code.

Also, properly refresh the page as `Ui.refresh` no longer exists.

Finally, previously, the refresh process was run along the token
request which means that we had a race condition that make the
refreshing of the page to had no effects. Now, we properly wrap
the token request with promises and close the window once all the
process is finished.